First functioning version of the program
The program in this commit receives a properties file, that contains the days limit, as input and
prints to the standard output the list of commits in the osquery repository owned by facebook. The
printed result is in JSON format and is displayed exactly how it is received from GitHub.
Currently the outputFile parameter in the properties file is not used